tests/requirements.txt
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
selenium==4.13.0
pytest-selenium==4.0.0
pytest-xdist==3.3.1
pytest-timeout==2.1.0
pytest-mock==3.11.1
httpx==0.25.0
aiohttp==3.9.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pytest-benchmark==3.4.1
pytest-asyncio==0.21.1
pytest-asyncio-engine==0.6.0
pytest-mock-resources==3.1.1
pytest-redis==2.0.0
pytest-postgresql==4.3.0
pytest-mysql==2.0.0
pytest-sqlalchemy==1.4.0
pytest-watch==4.2.0
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.6.1
pytest-mypy==0.10.0
